fix: Relocate classes to comply with CODING_RULE_STYLE_00009
This fix addresses CODING_RULE_STYLE_00009 violations by relocating
classes to their correct modules as specified in the AUTOSAR mapping.

## Changes

### 1. HwAttributeLiteralDef Relocation
- Moved from EcuResourceTemplate/HwAttributeValue.py to
  EcuResourceTemplate/HwElementCategory.py
- Deleted HwAttributeValue.py file (incorrectly named after a different
  class)

### 2. AtpBlueprintable Relocation
- Moved from GenericStructure/AbstractStructure.py to
  CommonStructure/StandardizationTemplate/AbstractBlueprintStructure/AtpBlueprint.py

## Files Modified
- 17 source files (class relocations and import updates)
- 2 test files (import updates)

## Test Coverage
All 255 tests for modified modules pass successfully.

Closes #378
