+++ src/armodel/models/M2/AUTOSARTemplates/CommonStructure/StandardizationTemplate/AbstractBlueprintStructure/AtpBlueprint.py
@@ -8,6 +8,29 @@ from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 
 
+class AtpBlueprintable(Identifiable, ABC):
+    """
+    Abstract base class for AUTOSAR Template (ATP) blueprintable elements.
+
+    AtpBlueprintable represents elements that can be used as blueprints in the AUTOSAR
+    template system. These elements provide reusable definitions that can be instantiated
+    or referenced in the model.
+
+    This class extends Identifiable with blueprint-specific functionality for managing
+    template-based elements in AUTOSAR models.
+
+    Note:
+        This is an abstract class and cannot be instantiated directly.
+        Concrete implementations include BswModuleEntry, CompuMethod, DataConstr,
+        and other blueprintable AUTOSAR elements.
+    """
+
+    def __init__(self, parent: ARObject, short_name: str):
+        if type(self) is AtpBlueprintable:
+            raise TypeError("AtpBlueprintable is an abstract class.")
+        super().__init__(parent, short_name)
+
+
 class AtpBlueprint(Identifiable, ABC):
     """
     Abstract base class for AUTOSAR Template (ATP) blueprint elements.
